`PhCalculation`: Fix bug when `only_initialization=True`
When the `only_initialization` setting is set to `True` the plugin would
add an instruction to the `remote_copy_list` to copy `DYN_MAT` folder
from the `parent_folder`. However, this would fail if the `parent_folder`
was that of a `PwCalculation`, which is one of the main use cases of the
`PhCalculation`.

It is not entirely clear why this was added in the commit that added it
dfdca7e1317484afedcf522b5adea17dc3f8533a which was just adding a plugin
for EPW. We assume it was added by mistake and so remove it.